+++ oc-includes/osclass/classes/datatables/MediaDataTable.php
@@ -20,6 +20,8 @@
  * @subpackage classes
  * @author     Shopclass
  */
+use mindstellar\admin\ListPaging;
+
 class MediaDataTable extends DataTable
 {
     private $order_by;
@@ -151,10 +153,8 @@ class MediaDataTable extends DataTable
         }
 
         // set start and limit using iPage param
-        $start = (Params::getParamInt('iPage') - 1) * $_get['iDisplayLength'];
-
-        $this->start = (int)$start;
-        $this->limit = (int)$_get['iDisplayLength'];
+        $this->limit = ListPaging::length((int)($_get['iDisplayLength'] ?? ListPaging::DEFAULT_LENGTH));
+        $this->start = ListPaging::start(ListPaging::page(), $this->limit);
     }
 
     /**