+++ oc-includes/osclass/classes/datatables/UsersDataTable.php
@@ -20,6 +20,8 @@
  * @subpackage classes
  * @author     Shopclass
  */
+use mindstellar\admin\ListPaging;
+
 class UsersDataTable extends DataTable
 {
     private $withUserId;
@@ -120,13 +122,7 @@ class UsersDataTable extends DataTable
             $_get['iDisplayStart'] = 0;
         }
         # Paging
-        $p_iPage = 1;
-        if (!is_numeric(Params::getParam('iPage')) || Params::getParam('iPage') < 1) {
-            Params::setParam('iPage', $p_iPage);
-            $this->iPage = $p_iPage;
-        } else {
-            $this->iPage = Params::getParam('iPage');
-        }
+        $this->iPage = ListPaging::page();
         # Sorting column and order
         $this->order_by = $this->resolveOrder($_get, $this->sortable, 'pk_i_id');
 
@@ -201,10 +197,8 @@ class UsersDataTable extends DataTable
         }
 
         // set start and limit using iPage param
-        $start = ($this->iPage - 1) * $_get['iDisplayLength'];
-
-        $this->start = (int)$start;
-        $this->limit = (int)$_get['iDisplayLength'];
+        $this->limit = ListPaging::length((int)($_get['iDisplayLength'] ?? ListPaging::DEFAULT_LENGTH));
+        $this->start = ListPaging::start($this->iPage, $this->limit);
     }
 
     /**